break cache.py into 4 files, fix a lot of patches in unittests
