+++ aws/logs_monitoring/tests/test_cloudtrail_s3.py
@@ -98,7 +98,7 @@ class TestS3CloudwatchParsing(unittest.TestCase):
     def setUp(self):
         self.maxDiff = 9000
 
-    @patch("cache.boto3")
+    @patch("base_tags_cache.boto3")
     @patch("parsing.boto3")
     @patch("lambda_function.boto3")
     def test_s3_cloudtrail_pasing_and_enrichment(
