+++ aws/logs_monitoring/tests/test_parsing.py
@@ -784,31 +784,29 @@ class TestParseSecurityHubEvents(unittest.TestCase):
 
 
 class TestAWSLogsHandler(unittest.TestCase):
-    @patch("cache.CloudwatchLogGroupTagsCache.release_s3_cache_lock")
-    @patch("cache.CloudwatchLogGroupTagsCache.acquire_s3_cache_lock")
-    @patch("cache.cloudwatch_logs_client")
-    @patch("cache.CloudwatchLogGroupTagsCache.write_cache_to_s3")
-    @patch("cache.send_forwarder_internal_metrics")
-    @patch("cache.CloudwatchLogGroupTagsCache.get_cache_from_s3")
+    @patch("parsing.CloudwatchLogGroupTagsCache.get")
+    @patch("parsing.CloudwatchLogGroupTagsCache.release_s3_cache_lock")
+    @patch("parsing.CloudwatchLogGroupTagsCache.acquire_s3_cache_lock")
+    @patch("parsing.CloudwatchLogGroupTagsCache.write_cache_to_s3")
+    @patch("base_tags_cache.send_forwarder_internal_metrics")
+    @patch("parsing.CloudwatchLogGroupTagsCache.get_cache_from_s3")
     def test_awslogs_handler_rds_postgresql(
         self,
         mock_get_s3_cache,
         mock_forward_metrics,
         mock_write_cache,
-        mock_boto3,
         mock_acquire_lock,
         mock_release_lock,
+        mock_cache_get,
     ):
         os.environ["DD_FETCH_LAMBDA_TAGS"] = "True"
         os.environ["DD_FETCH_LOG_GROUP_TAGS"] = "True"
         mock_acquire_lock.return_value = True
+        mock_cache_get.return_value = ["test_tag_key:test_tag_value"]
         mock_get_s3_cache.return_value = (
             {},
             1000,
         )
-        mock_boto3.list_tags_log_group.return_value = {
-            "tags": {"test_tag_key": "test_tag_value"}
-        }
 
         event = {
             "awslogs": {
@@ -862,6 +860,91 @@ class TestAWSLogsHandler(unittest.TestCase):
                 "ddtags": "env:dev,test_tag_key:test_tag_value,logname:postgresql",
                 "host": "datadog",
                 "service": "postgresql",
+            },
+            metadata,
+        )
+
+    @patch("parsing.CloudwatchLogGroupTagsCache.get")
+    @patch("parsing.StepFunctionsTagsCache.get")
+    @patch("parsing.StepFunctionsTagsCache.release_s3_cache_lock")
+    @patch("parsing.StepFunctionsTagsCache.acquire_s3_cache_lock")
+    @patch("parsing.StepFunctionsTagsCache.write_cache_to_s3")
+    @patch("base_tags_cache.send_forwarder_internal_metrics")
+    @patch("parsing.StepFunctionsTagsCache.get_cache_from_s3")
+    def test_awslogs_handler_step_functions_tags_added_properly(
+        self,
+        mock_get_s3_cache,
+        mock_forward_metrics,
+        mock_write_cache,
+        mock_acquire_lock,
+        mock_release_lock,
+        mock_step_functions_cache_get,
+        mock_cw_log_group_cache_get,
+    ):
+        os.environ["DD_FETCH_LAMBDA_TAGS"] = "True"
+        os.environ["DD_FETCH_LOG_GROUP_TAGS"] = "True"
+        os.environ["DD_FETCH_STEP_FUNCTIONS_TAGS"] = "True"
+        mock_acquire_lock.return_value = True
+        mock_step_functions_cache_get.return_value = ["test_tag_key:test_tag_value"]
+        mock_cw_log_group_cache_get.return_value = []
+        mock_get_s3_cache.return_value = (
+            {},
+            1000,
+        )
+
+        event = {
+            "awslogs": {
+                "data": base64.b64encode(
+                    gzip.compress(
+                        bytes(
+                            json.dumps(
+                                {
+                                    "messageType": "DATA_MESSAGE",
+                                    "owner": "425362996713",
+                                    "logGroup": "/aws/vendedlogs/states/logs-to-traces-sequential-Logs",
+                                    "logStream": "states/logs-to-traces-sequential/2022-11-10-15-50/7851b2d9",
+                                    "subscriptionFilters": ["testFilter"],
+                                    "logEvents": [
+                                        {
+                                            "id": "37199773595581154154810589279545129148442535997644275712",
+                                            "timestamp": 1668095539607,
+                                            "message": '{"id":"1","type":"ExecutionStarted","details":{"input":"{"Comment": "Insert your JSON here"}","inputDetails":{"truncated":false},"roleArn":"arn:aws:iam::425362996713:role/service-role/StepFunctions-logs-to-traces-sequential-role-ccd69c03"},",previous_event_id":"0","event_timestamp":"1668095539607","execution_arn":"arn:aws:states:sa-east-1:425362996713:express:logs-to-traces-sequential:d0dbefd8-a0f6-b402-da4c-f4863def7456:7fa0cfbe-be28-4a20-9875-73c37f5dc39e"}',
+                                        }
+                                    ],
+                                }
+                            ),
+                            "utf-8",
+                        )
+                    )
+                )
+            }
+        }
+        context = None
+        metadata = {"ddsource": "postgresql", "ddtags": "env:dev"}
+
+        self.assertEqual(
+            [
+                {
+                    "aws": {
+                        "awslogs": {
+                            "logGroup": "/aws/vendedlogs/states/logs-to-traces-sequential-Logs",
+                            "logStream": "states/logs-to-traces-sequential/2022-11-10-15-50/7851b2d9",
+                            "owner": "425362996713",
+                        }
+                    },
+                    "id": "37199773595581154154810589279545129148442535997644275712",
+                    "message": '{"id":"1","type":"ExecutionStarted","details":{"input":"{"Comment": "Insert your JSON here"}","inputDetails":{"truncated":false},"roleArn":"arn:aws:iam::425362996713:role/service-role/StepFunctions-logs-to-traces-sequential-role-ccd69c03"},",previous_event_id":"0","event_timestamp":"1668095539607","execution_arn":"arn:aws:states:sa-east-1:425362996713:express:logs-to-traces-sequential:d0dbefd8-a0f6-b402-da4c-f4863def7456:7fa0cfbe-be28-4a20-9875-73c37f5dc39e"}',
+                    "timestamp": 1668095539607,
+                }
+            ],
+            list(awslogs_handler(event, context, metadata)),
+        )
+        self.assertEqual(
+            {
+                "ddsource": "stepfunction",
+                "ddtags": "env:dev,test_tag_key:test_tag_value",
+                "host": "/aws/vendedlogs/states/logs-to-traces-sequential-Logs",
+                "service": "stepfunction",
             },
             metadata,
         )
